Refactoring of model and db code to be shared with api server.
