+++ backend/app/db/connection.py
@@ -1,24 +1,23 @@
-"""Database connection and session management for the Content Fetcher Service."""
+"""Database connection and engine management."""
 
 from sqlalchemy import create_engine, text
 from sqlalchemy.orm import declarative_base
-from sqlalchemy.orm import sessionmaker, Session
 from sqlalchemy.pool import StaticPool
-from typing import Generator
 import logging
-
-from .config import settings
 
 logger = logging.getLogger(__name__)
 
-# Create database engine with appropriate configuration
-def create_database_engine():
+# Base class for database models
+Base = declarative_base()
+
+
+def create_database_engine(database_url: str):
     """Create database engine with proper configuration for SQLite and PostgreSQL."""
     
-    if settings.database_url.startswith("sqlite"):
+    if database_url.startswith("sqlite"):
         # SQLite configuration for development
         engine = create_engine(
-            settings.database_url,
+            database_url,
             connect_args={"check_same_thread": False},
             poolclass=StaticPool,
             echo=False  # Set to True for SQL debugging
@@ -26,7 +25,7 @@ def create_database_engine():
     else:
         # PostgreSQL configuration for production
         engine = create_engine(
-            settings.database_url,
+            database_url,
             pool_size=5,
             max_overflow=10,
             pool_pre_ping=True,
@@ -35,33 +34,8 @@ def create_database_engine():
     
     return engine
 
-# Create engine and session factory
-engine = create_database_engine()
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-# Base class for database models
-Base = declarative_base()
-
-
-def get_database_session() -> Generator[Session, None, None]:
-    """
-    Create a database session for dependency injection.
-    
-    Yields:
-        Session: SQLAlchemy database session
-    """
-    session = SessionLocal()
-    try:
-        yield session
-    except Exception as e:
-        logger.error(f"Database session error: {e}")
-        session.rollback()
-        raise
-    finally:
-        session.close()
-
-
-def create_database_tables():
+def create_database_tables(engine):
     """Create all database tables. Used for initial setup."""
     try:
         Base.metadata.create_all(bind=engine)
@@ -71,10 +45,13 @@ def create_database_tables():
         raise
 
 
-def test_database_connection() -> bool:
+def test_database_connection(engine) -> bool:
     """
     Test database connectivity.
     
+    Args:
+        engine: SQLAlchemy engine
+        
     Returns:
         bool: True if connection successful, False otherwise
     """
@@ -85,4 +62,4 @@ def test_database_connection() -> bool:
         return True
     except Exception as e:
         logger.error(f"Database connection test failed: {e}")
-        return False
+        return False
