+++ backend/tests/fetcher/test_database.py
@@ -4,8 +4,7 @@ import pytest
 from unittest.mock import patch, MagicMock
 from sqlalchemy.exc import SQLAlchemyError
 
-from app.fetcher.database import (
-    create_database_engine, 
+from app.fetcher import (
     get_database_session, 
     create_database_tables,
     test_database_connection
@@ -18,23 +17,21 @@ class TestDatabaseEngine:
     
     def test_sqlite_engine_creation(self):
         """Test SQLite engine creation with proper configuration."""
-        with patch('app.fetcher.database.settings') as mock_settings:
-            mock_settings.database_url = "sqlite:///./test.db"
-            
-            engine = create_database_engine()
-            
-            assert engine is not None
-            assert "sqlite" in str(engine.url)
+        from app.db import create_database_engine
+        
+        engine = create_database_engine("sqlite:///./test.db")
+        
+        assert engine is not None
+        assert "sqlite" in str(engine.url)
     
     def test_postgresql_engine_creation(self):
         """Test PostgreSQL engine creation with proper configuration."""
-        with patch('app.fetcher.database.settings') as mock_settings:
-            mock_settings.database_url = "postgresql://user:pass@localhost/testdb"
-            
-            engine = create_database_engine()
-            
-            assert engine is not None
-            assert "postgresql" in str(engine.url)
+        from app.db import create_database_engine
+        
+        engine = create_database_engine("postgresql://user:pass@localhost/testdb")
+        
+        assert engine is not None
+        assert "postgresql" in str(engine.url)
 
 
 class TestDatabaseSession:
@@ -42,7 +39,7 @@ class TestDatabaseSession:
     
     def test_get_database_session(self):
         """Test database session creation and cleanup."""
-        with patch('app.fetcher.database.SessionLocal') as mock_session_local:
+        with patch('app.fetcher.SessionLocal') as mock_session_local:
             mock_session = MagicMock()
             mock_session_local.return_value = mock_session
             
@@ -55,7 +52,7 @@ class TestDatabaseSession:
     
     def test_get_database_session_with_exception(self):
         """Test database session cleanup on exception."""
-        with patch('app.fetcher.database.SessionLocal') as mock_session_local:
+        with patch('app.fetcher.SessionLocal') as mock_session_local:
             mock_session = MagicMock()
             mock_session_local.return_value = mock_session
             
@@ -80,8 +77,8 @@ class TestDatabaseSession:
 class TestDatabaseOperations:
     """Tests for database operations."""
     
-    @patch('app.fetcher.database.Base.metadata.create_all')
-    @patch('app.fetcher.database.engine')
+    @patch('app.fetcher.Base.metadata.create_all')
+    @patch('app.fetcher.engine')
     def test_create_database_tables_success(self, mock_engine, mock_create_all):
         """Test successful database table creation."""
         mock_create_all.return_value = None
@@ -91,8 +88,8 @@ class TestDatabaseOperations:
         
         mock_create_all.assert_called_once_with(bind=mock_engine)
     
-    @patch('app.fetcher.database.Base.metadata.create_all')
-    @patch('app.fetcher.database.engine')
+    @patch('app.fetcher.Base.metadata.create_all')
+    @patch('app.fetcher.engine')
     def test_create_database_tables_failure(self, mock_engine, mock_create_all):
         """Test database table creation failure."""
         mock_create_all.side_effect = SQLAlchemyError("Table creation failed")
@@ -100,7 +97,7 @@ class TestDatabaseOperations:
         with pytest.raises(SQLAlchemyError):
             create_database_tables()
     
-    @patch('app.fetcher.database.engine')
+    @patch('app.fetcher.engine')
     def test_database_connection_success(self, mock_engine):
         """Test successful database connection test."""
         mock_connection = MagicMock()
@@ -112,7 +109,7 @@ class TestDatabaseOperations:
         # Verify execute was called (the text() wrapper will be used internally)
         mock_connection.execute.assert_called_once()
     
-    @patch('app.fetcher.database.engine')
+    @patch('app.fetcher.engine')
     def test_database_connection_failure(self, mock_engine):
         """Test database connection test failure."""
         mock_engine.connect.side_effect = SQLAlchemyError("Connection failed")
